Домерджил с мастером.

+++ base.cpp
@@ -22,7 +22,7 @@ loadConfig(GameState *state, const std::string& configPath, Config *config)
     config->defaultScene = file.parseElement<std::string>("scene", "default");
 
     Parser opKeys(configPath,{"input", "oppositeKeys"});
-    for (int i = 0; i < opKeys.size(); ++i)
+    for (int i = 0; i < (int) opKeys.size(); ++i)
     {
         std::string key1 = opKeys.parseObjectElement<std::string>(i, 0);
         std::string key2 = opKeys.parseObjectElement<std::string>(i,  1);
@@ -30,7 +30,7 @@ loadConfig(GameState *state, const std::string& configPath, Config *config)
     }
 
     Parser mainKeys(configPath,{"input", "keys"});
-    for (int i = 0; i < mainKeys.size(); ++i)
+    for (int i = 0; i < (int) mainKeys.size(); ++i)
     {
         std::string key = mainKeys.parseObjectElement<std::string>(i,0);
         std::string value = mainKeys.parseObjectElement<std::string>(i, 1);
@@ -38,7 +38,7 @@ loadConfig(GameState *state, const std::string& configPath, Config *config)
     }
 
     Parser axisKeys(configPath,{"input", "axisData"});
-    for (int i = 0; i < axisKeys.size(); ++i)
+    for (int i = 0; i < (int) axisKeys.size(); ++i)
     {
        KeyData data;
 
@@ -52,7 +52,7 @@ loadConfig(GameState *state, const std::string& configPath, Config *config)
     }
 
     Parser axes(configPath,{"input","axes"});
-    for (int i = 0; i < axes.size(); ++i)
+    for (int i = 0; i < (int) axes.size(); ++i)
     {
         std::string axis = axes.parseObjectElement<std::string>(i, 0);
         auto value = axes.parseObjectElement<float>(i, 1);
